refactor: SearchResults 페이지 라우팅 구조 변경 - nested 라우팅으로 layout 공유 - 경로 변경
