+++ styles/pages/searchResults/companyListWrapper.ts
