+++ styles/pages/searchResults/personListWrapper.ts
